tests/integration/requirements.txt
pytest>=8.0
pyyaml>=6.0
requests>=2.31
pytest-xdist>=3.5.0
